+++ packages/client/src/tx/signed-extensions/chain/CheckTxVersion.ts
@@ -1,8 +1,8 @@
 import { map } from "rxjs"
 import type { GetChainSignedExtension } from "../internal-types"
-import { empty, systemVersionProp$ } from "../utils"
+import { additionalSigned, systemVersionProp$ } from "../utils"
 
 export const CheckTxVersion: GetChainSignedExtension = ({ lookupFn }) =>
   systemVersionProp$("transaction_version", lookupFn).pipe(
-    map((additionalSigned) => ({ additionalSigned, value: empty })),
+    map(additionalSigned),
   )