+++ packages/observable-client/src/chainHead/chainHead.ts
@@ -13,6 +13,7 @@ import {
   Subject,
   defer,
   distinctUntilChanged,
+  filter,
   map,
   merge,
   mergeMap,
@@ -46,6 +47,7 @@ import type {
 import { getFollow$, getPinnedBlocks$ } from "./streams"
 import { getTrackTx } from "./track-tx"
 import { getValidateTx } from "./validate-tx"
+import { HexString } from "@polkadot-api/substrate-bindings"
 
 export type {
   PinnedBlocks,
@@ -400,6 +402,37 @@ export const getChainHead$ = (chainHead: ChainHead) => {
     unfollow = startFollow()
   }
 
+  const genesis$ = runtime$.pipe(
+    filter(Boolean),
+    take(1),
+    mergeMap((runtime) => {
+      const { enc } = runtime.dynamicBuilder.buildStorage(
+        "System",
+        "BlockHash",
+      ).keys
+      // const genesis$ =
+      // there are chains (e.g. kilt) that use u64 as block number
+      // u64 is encoded as bigint
+      // using dynamic builder for safety
+      let key: string
+      try {
+        // for u32
+        key = enc(0)
+      } catch {
+        // for u64
+        key = enc(0n)
+      }
+
+      return storage$(
+        null,
+        false,
+        "value",
+        () => key,
+        null,
+      ) as Observable<HexString>
+    }),
+  )
+
   return [
     {
       follow$,
@@ -408,6 +441,7 @@ export const getChainHead$ = (chainHead: ChainHead) => {
       bestBlocks$,
       runtime$,
       metadata$,
+      genesis$,
 
       header$,
       body$,